extract e2e logic in lib
