+++ ceno_zkvm/src/lib.rs
@@ -11,6 +11,7 @@ pub mod tables;
 pub use utils::u64vec;
 mod chip_handler;
 pub mod circuit_builder;
+pub mod e2e;
 pub mod expression;
 pub mod gadgets;
 mod keygen;
